Run queue events ASAP (#186)
It turns out that waiting for `analytics.ready` can create a delay
in when automatic `page` calls will be fired.

This is due to some destinations only ever being ready once two
page calls have been fired. These destinations are the ones that
have the `_assumePageView` property set to true.