persisted.ts : Strengthen conditional checking for localStorage to prevent "null" from being used
DOM storage can be disabled for the Android WebView which results in "window.localStorage" being "null". Because of this we are adding an extra condition to check for the existence of "window.localStorage". If not truthy, then ternary will inevitably return the default "loc" object as the stubbed implementation for "localStorage".