Batched dispatcher fixes and enhancements
Fixes event delivery issue when `analytics.track` is called right before
an action that causes page to unload, for example clicking a link or closing
a tab. The problem was caused by `beforeunload` handler being called before
`dispatch` which actually pushes the event to the buffer, since latter is part
of an async pipeline, unlike `beforeunload` sequence which is synchronous. The
issue is fixed by adding a `pageUnloaded` flag which gets set to true in the
`beforeunload` handler, so if `dispatch` gets called any time after that, it'd
know to send the event right away since there's no time to waste as page is about
to be unloaded.

Enhancements:

- Removed the `flushing` flag. It's value is changed from `true` to `false` synchronously
inside `flush`, so effectively the value of `flushing` is always `false` therefore serving
no actual purpose.

- Simplified the `buffer` type. The first member of the inner array serves no purpose, so
it's been flattened to simple 1D array.

- De-duplicated `fetch` calls into a common `sendBatch` function.

- Flush scheduler is now self-contained. Previoulsy the timer ID was managed by callers,
which means there was a potential for a caller to not check if flush has already been
scheduled. This could've led to "phantom" timers if `scheduleFlush` had been called multiple
times without checking.

- Window unload tests are now more realistic as they *trigger* the `beforeunload` event instead
calling the handler directly.